#32 merged image and text creation into one view. fixed permission bug with camera. small improvements to create task view design.
