+++ App.js
@@ -2,8 +2,7 @@ import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import {createStackNavigator} from "react-navigation";
 import HomeView from "./components/HomeViewComponent/HomeView";
-import CreateTextTaskView from "./components/CreateTextTaskViewComponent/CreateTextTaskView";
-import CreateImageTaskView from "./components/CreateImageTaskViewComponent/CreateImageTaskView";
+import CreateTaskView from "./components/CreateTaskViewComponent/CreateTaskView";
 
 export default class App extends React.Component {
 
@@ -15,8 +14,7 @@ export default class App extends React.Component {
 const AppWithNavigation = createStackNavigator(
     {
         Home: HomeView,
-        TextTask: CreateTextTaskView,
-        ImageTask: CreateImageTaskView,
+        NewTask: CreateTaskView,
     },
     {
         initialRouteName: 'Home',
