+++ components/ChooseImageComponent/styles/styles.js
@@ -1,0 +1,29 @@
+import { StyleSheet } from 'react-native';
+
+
+export default StyleSheet.create({
+
+    image: {
+        borderColor: '#d6d6d6',
+        borderWidth: 2,
+        borderRadius: 10,
+        width: '90%',
+        aspectRatio: 1,
+        backgroundColor: '#e8e8e8',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+
+    imageContainer: {
+        marginTop: 20,
+        width: '80%',
+        aspectRatio: 1,
+        alignItems: 'center',
+    },
+
+    text: {
+        fontSize: 24,
+        color: '#d6d6d6',
+    },
+})
