+++ components/ChooseTextComponent/styles/styles.js
@@ -1,0 +1,17 @@
+import {StyleSheet} from 'react-native';
+
+export default StyleSheet.create({
+    textInputField:{
+        height: 200,
+        borderRadius: 15,
+        borderWidth: 1,
+        borderColor: '#f3f3f3',
+        backgroundColor:'#f3f3f3',
+        padding:10,
+        width: "90%",
+        marginTop:20,
+        marginBottom:20,
+        fontSize: 17,
+        fontWeight:"bold",
+    },
+})
