+++ components/UploadImageModalComponent/styles/styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+
+export default StyleSheet.create({
+    modal: {
+        backgroundColor: 'rgba(255,255,255,0)',
+        width: '90%',
+        display: 'flex',
+        justifyContent: 'flex-end',
+    },
+
+    button: {
+        backgroundColor: '#ff0042',
+        borderRadius: 10,
+        margin: 1,
+        width: '100%',
+    },
+})
